fix(fontAweseme): added font Awesome
refactor code abit tried to move some code for fontawesome.
