+++ FontAwesome/FontAwesomeHelper.cs
@@ -1,0 +1,1652 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Drawing;
+using System.Drawing.Text;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Reflection;
+using System.IO;
+using System.Runtime.InteropServices;
+
+
+//CONTINUE THIS IMPLEMENTATION!
+//https://stackoverflow.com/questions/556147/how-do-i-embed-my-own-fonts-in-a-winforms-app
+
+namespace FontAwesome
+{
+    public class FontAwesomeHelper
+    {
+        //[System.Runtime.InteropServices.DllImport("gdi32.dll")]
+        //private static extern IntPtr AddFontMemResourceEx(IntPtr pbFont, uint cbFont, IntPtr pdv, [System.Runtime.InteropServices.In] ref uint pcFonts);
+
+        private PrivateFontCollection _fonts;
+
+        public FontAwesomeHelper()
+        {
+            //Resource.fontawesome_webfont
+            try
+            {
+                _fonts = new PrivateFontCollection();
+                ////_fonts.AddFontFile(@".\Resources\fontawesome-webfont.ttf");
+
+                //byte[] fontData = Resource.fontawesome_webfont;
+                //IntPtr fontPtr = System.Runtime.InteropServices.Marshal.AllocCoTaskMem(fontData.Length);
+                //System.Runtime.InteropServices.Marshal.Copy(fontData, 0, fontPtr, fontData.Length);
+                //uint dummy = 0;
+                //_fonts.AddMemoryFont(fontPtr, Resource.fontawesome_webfont.Length);
+                //AddFontMemResourceEx(fontPtr, (uint)Resource.fontawesome_webfont.Length, IntPtr.Zero, ref dummy);
+                //System.Runtime.InteropServices.Marshal.FreeCoTaskMem(fontPtr);
+
+                ////var myFont = new Font(_fonts.Families[0], 16.0F);
+
+
+                //using (Stream fontStream = Assembly.GetExecutingAssembly().GetManifestResourceStream("Resources/fontawesome-webfont.ttf"))
+                //{
+                    byte[] fontdata = Resource.fontawesome_webfont;
+                    var data = Marshal.AllocCoTaskMem(fontdata.Length);
+                    //byte[] fontdata = new byte[fontStream.Length];
+                    //fontStream.Read(fontdata, 0, (int)fontStream.Length);
+                    Marshal.Copy(fontdata, 0, data, fontdata.Length);
+                    _fonts.AddMemoryFont(data, fontdata.Length);
+                    //fontStream.Close();
+                    Marshal.FreeCoTaskMem(data);
+                //}
+
+            }
+            catch (System.Exception)
+            {
+                //Just ingore if we get problems with font loading!
+            }
+        }
+
+        public System.Drawing.Font GetFont(int size = 16)
+        {
+            //Fallback if we fail to load font
+            if (_fonts.Families.Length == 0) return System.Drawing.SystemFonts.DefaultFont;
+
+            return new System.Drawing.Font(_fonts.Families[0], size, System.Drawing.FontStyle.Regular);
+        }
+
+        public string LoadIcon(string input)
+        {
+            switch (input)
+            {
+                case "fa-500px":
+                    return "\xf26e";
+                case "fa-address-book":
+                    return "\xf2b9";
+                case "fa-address-book-o":
+                    return "\xf2ba";
+                case "fa-address-card":
+                    return "\xf2bb";
+                case "fa-address-card-o":
+                    return "\xf2bc";
+                case "fa-adjust":
+                    return "\xf042";
+                case "fa-adn":
+                    return "\xf170";
+                case "fa-align-center":
+                    return "\xf037";
+                case "fa-align-justify":
+                    return "\xf039";
+                case "fa-align-left":
+                    return "\xf036";
+                case "fa-align-right":
+                    return "\xf038";
+                case "fa-amazon":
+                    return "\xf270";
+                case "fa-ambulance":
+                    return "\xf0f9";
+                case "fa-american-sign-language-interpreting":
+                    return "\xf2a3";
+                case "fa-anchor":
+                    return "\xf13d";
+                case "fa-android":
+                    return "\xf17b";
+                case "fa-angellist":
+                    return "\xf209";
+                case "fa-angle-double-down":
+                    return "\xf103";
+                case "fa-angle-double-left":
+                    return "\xf100";
+                case "fa-angle-double-right":
+                    return "\xf101";
+                case "fa-angle-double-up":
+                    return "\xf102";
+                case "fa-angle-down":
+                    return "\xf107";
+                case "fa-angle-left":
+                    return "\xf104";
+                case "fa-angle-right":
+                    return "\xf105";
+                case "fa-angle-up":
+                    return "\xf106";
+                case "fa-apple":
+                    return "\xf179";
+                case "fa-archive":
+                    return "\xf187";
+                case "fa-area-chart":
+                    return "\xf1fe";
+                case "fa-arrow-circle-down":
+                    return "\xf0ab";
+                case "fa-arrow-circle-left":
+                    return "\xf0a8";
+                case "fa-arrow-circle-o-down":
+                    return "\xf01a";
+                case "fa-arrow-circle-o-left":
+                    return "\xf190";
+                case "fa-arrow-circle-o-right":
+                    return "\xf18e";
+                case "fa-arrow-circle-o-up":
+                    return "\xf01b";
+                case "fa-arrow-circle-right":
+                    return "\xf0a9";
+                case "fa-arrow-circle-up":
+                    return "\xf0aa";
+                case "fa-arrow-down":
+                    return "\xf063";
+                case "fa-arrow-left":
+                    return "\xf060";
+                case "fa-arrow-right":
+                    return "\xf061";
+                case "fa-arrow-up":
+                    return "\xf062";
+                case "fa-arrows":
+                    return "\xf047";
+                case "fa-arrows-alt":
+                    return "\xf0b2";
+                case "fa-arrows-h":
+                    return "\xf07e";
+                case "fa-arrows-v":
+                    return "\xf07d";
+                case "fa-asl-interpreting":
+                    return "\xf2a3";
+                case "fa-assistive-listening-systems":
+                    return "\xf2a2";
+                case "fa-asterisk":
+                    return "\xf069";
+                case "fa-at":
+                    return "\xf1fa";
+                case "fa-audio-description":
+                    return "\xf29e";
+                case "fa-automobile":
+                    return "\xf1b9";
+                case "fa-backward":
+                    return "\xf04a";
+                case "fa-balance-scale":
+                    return "\xf24e";
+                case "fa-ban":
+                    return "\xf05e";
+                case "fa-bandcamp":
+                    return "\xf2d5";
+                case "fa-bank":
+                    return "\xf19c";
+                case "fa-bar-chart":
+                    return "\xf080";
+                case "fa-bar-chart-o":
+                    return "\xf080";
+                case "fa-barcode":
+                    return "\xf02a";
+                case "fa-bars":
+                    return "\xf0c9";
+                case "fa-bath":
+                    return "\xf2cd";
+                case "fa-bathtub":
+                    return "\xf2cd";
+                case "fa-battery":
+                    return "\xf240";
+                case "fa-battery-0":
+                    return "\xf244";
+                case "fa-battery-1":
+                    return "\xf243";
+                case "fa-battery-2":
+                    return "\xf242";
+                case "fa-battery-3":
+                    return "\xf241";
+                case "fa-battery-4":
+                    return "\xf240";
+                case "fa-battery-empty":
+                    return "\xf244";
+                case "fa-battery-full":
+                    return "\xf240";
+                case "fa-battery-half":
+                    return "\xf242";
+                case "fa-battery-quarter":
+                    return "\xf243";
+                case "fa-battery-three-quarters":
+                    return "\xf241";
+                case "fa-bed":
+                    return "\xf236";
+                case "fa-beer":
+                    return "\xf0fc";
+                case "fa-behance":
+                    return "\xf1b4";
+                case "fa-behance-square":
+                    return "\xf1b5";
+                case "fa-bell":
+                    return "\xf0f3";
+                case "fa-bell-o":
+                    return "\xf0a2";
+                case "fa-bell-slash":
+                    return "\xf1f6";
+                case "fa-bell-slash-o":
+                    return "\xf1f7";
+                case "fa-bicycle":
+                    return "\xf206";
+                case "fa-binoculars":
+                    return "\xf1e5";
+                case "fa-birthday-cake":
+                    return "\xf1fd";
+                case "fa-bitbucket":
+                    return "\xf171";
+                case "fa-bitbucket-square":
+                    return "\xf172";
+                case "fa-bitcoin":
+                    return "\xf15a";
+                case "fa-black-tie":
+                    return "\xf27e";
+                case "fa-blind":
+                    return "\xf29d";
+                case "fa-bluetooth":
+                    return "\xf293";
+                case "fa-bluetooth-b":
+                    return "\xf294";
+                case "fa-bold":
+                    return "\xf032";
+                case "fa-bolt":
+                    return "\xf0e7";
+                case "fa-bomb":
+                    return "\xf1e2";
+                case "fa-book":
+                    return "\xf02d";
+                case "fa-bookmark":
+                    return "\xf02e";
+                case "fa-bookmark-o":
+                    return "\xf097";
+                case "fa-braille":
+                    return "\xf2a1";
+                case "fa-briefcase":
+                    return "\xf0b1";
+                case "fa-btc":
+                    return "\xf15a";
+                case "fa-bug":
+                    return "\xf188";
+                case "fa-building":
+                    return "\xf1ad";
+                case "fa-building-o":
+                    return "\xf0f7";
+                case "fa-bullhorn":
+                    return "\xf0a1";
+                case "fa-bullseye":
+                    return "\xf140";
+                case "fa-bus":
+                    return "\xf207";
+                case "fa-buysellads":
+                    return "\xf20d";
+                case "fa-cab":
+                    return "\xf1ba";
+                case "fa-calculator":
+                    return "\xf1ec";
+                case "fa-calendar":
+                    return "\xf073";
+                case "fa-calendar-check-o":
+                    return "\xf274";
+                case "fa-calendar-minus-o":
+                    return "\xf272";
+                case "fa-calendar-o":
+                    return "\xf133";
+                case "fa-calendar-plus-o":
+                    return "\xf271";
+                case "fa-calendar-times-o":
+                    return "\xf273";
+                case "fa-camera":
+                    return "\xf030";
+                case "fa-camera-retro":
+                    return "\xf083";
+                case "fa-car":
+                    return "\xf1b9";
+                case "fa-caret-down":
+                    return "\xf0d7";
+                case "fa-caret-left":
+                    return "\xf0d9";
+                case "fa-caret-right":
+                    return "\xf0da";
+                case "fa-caret-square-o-down":
+                    return "\xf150";
+                case "fa-caret-square-o-left":
+                    return "\xf191";
+                case "fa-caret-square-o-right":
+                    return "\xf152";
+                case "fa-caret-square-o-up":
+                    return "\xf151";
+                case "fa-caret-up":
+                    return "\xf0d8";
+                case "fa-cart-arrow-down":
+                    return "\xf218";
+                case "fa-cart-plus":
+                    return "\xf217";
+                case "fa-cc":
+                    return "\xf20a";
+                case "fa-cc-amex":
+                    return "\xf1f3";
+                case "fa-cc-diners-club":
+                    return "\xf24c";
+                case "fa-cc-discover":
+                    return "\xf1f2";
+                case "fa-cc-jcb":
+                    return "\xf24b";
+                case "fa-cc-mastercard":
+                    return "\xf1f1";
+                case "fa-cc-paypal":
+                    return "\xf1f4";
+                case "fa-cc-stripe":
+                    return "\xf1f5";
+                case "fa-cc-visa":
+                    return "\xf1f0";
+                case "fa-certificate":
+                    return "\xf0a3";
+                case "fa-chain":
+                    return "\xf0c1";
+                case "fa-chain-broken":
+                    return "\xf127";
+                case "fa-check":
+                    return "\xf00c";
+                case "fa-check-circle":
+                    return "\xf058";
+                case "fa-check-circle-o":
+                    return "\xf05d";
+                case "fa-check-square":
+                    return "\xf14a";
+                case "fa-check-square-o":
+                    return "\xf046";
+                case "fa-chevron-circle-down":
+                    return "\xf13a";
+                case "fa-chevron-circle-left":
+                    return "\xf137";
+                case "fa-chevron-circle-right":
+                    return "\xf138";
+                case "fa-chevron-circle-up":
+                    return "\xf139";
+                case "fa-chevron-down":
+                    return "\xf078";
+                case "fa-chevron-left":
+                    return "\xf053";
+                case "fa-chevron-right":
+                    return "\xf054";
+                case "fa-chevron-up":
+                    return "\xf077";
+                case "fa-child":
+                    return "\xf1ae";
+                case "fa-chrome":
+                    return "\xf268";
+                case "fa-circle":
+                    return "\xf111";
+                case "fa-circle-o":
+                    return "\xf10c";
+                case "fa-circle-o-notch":
+                    return "\xf1ce";
+                case "fa-circle-thin":
+                    return "\xf1db";
+                case "fa-clipboard":
+                    return "\xf0ea";
+                case "fa-clock-o":
+                    return "\xf017";
+                case "fa-clone":
+                    return "\xf24d";
+                case "fa-close":
+                    return "\xf00d";
+                case "fa-cloud":
+                    return "\xf0c2";
+                case "fa-cloud-download":
+                    return "\xf0ed";
+                case "fa-cloud-upload":
+                    return "\xf0ee";
+                case "fa-cny":
+                    return "\xf157";
+                case "fa-code":
+                    return "\xf121";
+                case "fa-code-fork":
+                    return "\xf126";
+                case "fa-codepen":
+                    return "\xf1cb";
+                case "fa-codiepie":
+                    return "\xf284";
+                case "fa-coffee":
+                    return "\xf0f4";
+                case "fa-cog":
+                    return "\xf013";
+                case "fa-cogs":
+                    return "\xf085";
+                case "fa-columns":
+                    return "\xf0db";
+                case "fa-comment":
+                    return "\xf075";
+                case "fa-comment-o":
+                    return "\xf0e5";
+                case "fa-commenting":
+                    return "\xf27a";
+                case "fa-commenting-o":
+                    return "\xf27b";
+                case "fa-comments":
+                    return "\xf086";
+                case "fa-comments-o":
+                    return "\xf0e6";
+                case "fa-compass":
+                    return "\xf14e";
+                case "fa-compress":
+                    return "\xf066";
+                case "fa-connectdevelop":
+                    return "\xf20e";
+                case "fa-contao":
+                    return "\xf26d";
+                case "fa-copy":
+                    return "\xf0c5";
+                case "fa-copyright":
+                    return "\xf1f9";
+                case "fa-creative-commons":
+                    return "\xf25e";
+                case "fa-credit-card":
+                    return "\xf09d";
+                case "fa-credit-card-alt":
+                    return "\xf283";
+                case "fa-crop":
+                    return "\xf125";
+                case "fa-crosshairs":
+                    return "\xf05b";
+                case "fa-css3":
+                    return "\xf13c";
+                case "fa-cube":
+                    return "\xf1b2";
+                case "fa-cubes":
+                    return "\xf1b3";
+                case "fa-cut":
+                    return "\xf0c4";
+                case "fa-cutlery":
+                    return "\xf0f5";
+                case "fa-dashboard":
+                    return "\xf0e4";
+                case "fa-dashcube":
+                    return "\xf210";
+                case "fa-database":
+                    return "\xf1c0";
+                case "fa-deaf":
+                    return "\xf2a4";
+                case "fa-deafness":
+                    return "\xf2a4";
+                case "fa-dedent":
+                    return "\xf03b";
+                case "fa-delicious":
+                    return "\xf1a5";
+                case "fa-desktop":
+                    return "\xf108";
+                case "fa-deviantart":
+                    return "\xf1bd";
+                case "fa-diamond":
+                    return "\xf219";
+                case "fa-digg":
+                    return "\xf1a6";
+                case "fa-dollar":
+                    return "\xf155";
+                case "fa-dot-circle-o":
+                    return "\xf192";
+                case "fa-download":
+                    return "\xf019";
+                case "fa-dribbble":
+                    return "\xf17d";
+                case "fa-drivers-license":
+                    return "\xf2c2";
+                case "fa-drivers-license-o":
+                    return "\xf2c3";
+                case "fa-dropbox":
+                    return "\xf16b";
+                case "fa-drupal":
+                    return "\xf1a9";
+                case "fa-edge":
+                    return "\xf282";
+                case "fa-edit":
+                    return "\xf044";
+                case "fa-eercast":
+                    return "\xf2da";
+                case "fa-eject":
+                    return "\xf052";
+                case "fa-ellipsis-h":
+                    return "\xf141";
+                case "fa-ellipsis-v":
+                    return "\xf142";
+                case "fa-empire":
+                    return "\xf1d1";
+                case "fa-envelope":
+                    return "\xf0e0";
+                case "fa-envelope-o":
+                    return "\xf003";
+                case "fa-envelope-open":
+                    return "\xf2b6";
+                case "fa-envelope-open-o":
+                    return "\xf2b7";
+                case "fa-envelope-square":
+                    return "\xf199";
+                case "fa-envira":
+                    return "\xf299";
+                case "fa-eraser":
+                    return "\xf12d";
+                case "fa-etsy":
+                    return "\xf2d7";
+                case "fa-eur":
+                    return "\xf153";
+                case "fa-euro":
+                    return "\xf153";
+                case "fa-exchange":
+                    return "\xf0ec";
+                case "fa-exclamation":
+                    return "\xf12a";
+                case "fa-exclamation-circle":
+                    return "\xf06a";
+                case "fa-exclamation-triangle":
+                    return "\xf071";
+                case "fa-expand":
+                    return "\xf065";
+                case "fa-expeditedssl":
+                    return "\xf23e";
+                case "fa-external-link":
+                    return "\xf08e";
+                case "fa-external-link-square":
+                    return "\xf14c";
+                case "fa-eye":
+                    return "\xf06e";
+                case "fa-eye-slash":
+                    return "\xf070";
+                case "fa-eyedropper":
+                    return "\xf1fb";
+                case "fa-fa":
+                    return "\xf2b4";
+                case "fa-facebook":
+                    return "\xf09a";
+                case "fa-facebook-f":
+                    return "\xf09a";
+                case "fa-facebook-official":
+                    return "\xf230";
+                case "fa-facebook-square":
+                    return "\xf082";
+                case "fa-fast-backward":
+                    return "\xf049";
+                case "fa-fast-forward":
+                    return "\xf050";
+                case "fa-fax":
+                    return "\xf1ac";
+                case "fa-feed":
+                    return "\xf09e";
+                case "fa-female":
+                    return "\xf182";
+                case "fa-fighter-jet":
+                    return "\xf0fb";
+                case "fa-file":
+                    return "\xf15b";
+                case "fa-file-archive-o":
+                    return "\xf1c6";
+                case "fa-file-audio-o":
+                    return "\xf1c7";
+                case "fa-file-code-o":
+                    return "\xf1c9";
+                case "fa-file-excel-o":
+                    return "\xf1c3";
+                case "fa-file-image-o":
+                    return "\xf1c5";
+                case "fa-file-movie-o":
+                    return "\xf1c8";
+                case "fa-file-o":
+                    return "\xf016";
+                case "fa-file-pdf-o":
+                    return "\xf1c1";
+                case "fa-file-photo-o":
+                    return "\xf1c5";
+                case "fa-file-picture-o":
+                    return "\xf1c5";
+                case "fa-file-powerpoint-o":
+                    return "\xf1c4";
+                case "fa-file-sound-o":
+                    return "\xf1c7";
+                case "fa-file-text":
+                    return "\xf15c";
+                case "fa-file-text-o":
+                    return "\xf0f6";
+                case "fa-file-video-o":
+                    return "\xf1c8";
+                case "fa-file-word-o":
+                    return "\xf1c2";
+                case "fa-file-zip-o":
+                    return "\xf1c6";
+                case "fa-files-o":
+                    return "\xf0c5";
+                case "fa-film":
+                    return "\xf008";
+                case "fa-filter":
+                    return "\xf0b0";
+                case "fa-fire":
+                    return "\xf06d";
+                case "fa-fire-extinguisher":
+                    return "\xf134";
+                case "fa-firefox":
+                    return "\xf269";
+                case "fa-first-order":
+                    return "\xf2b0";
+                case "fa-flag":
+                    return "\xf024";
+                case "fa-flag-checkered":
+                    return "\xf11e";
+                case "fa-flag-o":
+                    return "\xf11d";
+                case "fa-flash":
+                    return "\xf0e7";
+                case "fa-flask":
+                    return "\xf0c3";
+                case "fa-flickr":
+                    return "\xf16e";
+                case "fa-floppy-o":
+                    return "\xf0c7";
+                case "fa-folder":
+                    return "\xf07b";
+                case "fa-folder-o":
+                    return "\xf114";
+                case "fa-folder-open":
+                    return "\xf07c";
+                case "fa-folder-open-o":
+                    return "\xf115";
+                case "fa-font":
+                    return "\xf031";
+                case "fa-font-awesome":
+                    return "\xf2b4";
+                case "fa-fonticons":
+                    return "\xf280";
+                case "fa-fort-awesome":
+                    return "\xf286";
+                case "fa-forumbee":
+                    return "\xf211";
+                case "fa-forward":
+                    return "\xf04e";
+                case "fa-foursquare":
+                    return "\xf180";
+                case "fa-free-code-camp":
+                    return "\xf2c5";
+                case "fa-frown-o":
+                    return "\xf119";
+                case "fa-futbol-o":
+                    return "\xf1e3";
+                case "fa-gamepad":
+                    return "\xf11b";
+                case "fa-gavel":
+                    return "\xf0e3";
+                case "fa-gbp":
+                    return "\xf154";
+                case "fa-ge":
+                    return "\xf1d1";
+                case "fa-gear":
+                    return "\xf013";
+                case "fa-gears":
+                    return "\xf085";
+                case "fa-genderless":
+                    return "\xf22d";
+                case "fa-get-pocket":
+                    return "\xf265";
+                case "fa-gg":
+                    return "\xf260";
+                case "fa-gg-circle":
+                    return "\xf261";
+                case "fa-gift":
+                    return "\xf06b";
+                case "fa-git":
+                    return "\xf1d3";
+                case "fa-git-square":
+                    return "\xf1d2";
+                case "fa-github":
+                    return "\xf09b";
+                case "fa-github-alt":
+                    return "\xf113";
+                case "fa-github-square":
+                    return "\xf092";
+                case "fa-gitlab":
+                    return "\xf296";
+                case "fa-gittip":
+                    return "\xf184";
+                case "fa-glass":
+                    return "\xf000";
+                case "fa-glide":
+                    return "\xf2a5";
+                case "fa-glide-g":
+                    return "\xf2a6";
+                case "fa-globe":
+                    return "\xf0ac";
+                case "fa-google":
+                    return "\xf1a0";
+                case "fa-google-plus":
+                    return "\xf0d5";
+                case "fa-google-plus-circle":
+                    return "\xf2b3";
+                case "fa-google-plus-official":
+                    return "\xf2b3";
+                case "fa-google-plus-square":
+                    return "\xf0d4";
+                case "fa-google-wallet":
+                    return "\xf1ee";
+                case "fa-graduation-cap":
+                    return "\xf19d";
+                case "fa-gratipay":
+                    return "\xf184";
+                case "-grav":
+                    return "\xf2d6";
+                case "-group":
+                    return "\xf0c0";
+                case "fa-h-square":
+                    return "\xf0fd";
+                case "fa-hacker-news":
+                    return "\xf1d4";
+                case "fa-hand-grab-o":
+                    return "\xf255";
+                case "fa-hand-lizard-o":
+                    return "\xf258";
+                case "fa-hand-o-down":
+                    return "\xf0a7";
+                case "fa-hand-o-left":
+                    return "\xf0a5";
+                case "fa-hand-o-right":
+                    return "\xf0a4";
+                case "fa-hand-o-up":
+                    return "\xf0a6";
+                case "fa-hand-paper-o":
+                    return "\xf256";
+                case "fa-hand-peace-o":
+                    return "\xf25b";
+                case "fa-hand-pointer-o":
+                    return "\xf25a";
+                case "fa-hand-rock-o":
+                    return "\xf255";
+                case "fa-hand-scissors-o":
+                    return "\xf257";
+                case "fa-hand-spock-o":
+                    return "\xf259";
+                case "fa-hand-stop-o":
+                    return "\xf256";
+                case "fa-handshake-o":
+                    return "\xf2b5";
+                case "fa-hard-of-hearing":
+                    return "\xf2a4";
+                case "fa-hashtag":
+                    return "\xf292";
+                case "fa-hdd-o":
+                    return "\xf0a0";
+                case "fa-header":
+                    return "\xf1dc";
+                case "fa-headphones":
+                    return "\xf025";
+                case "fa-heart":
+                    return "\xf004";
+                case "fa-heart-o":
+                    return "\xf08a";
+                case "fa-heartbeat":
+                    return "\xf21e";
+                case "fa-history":
+                    return "\xf1da";
+                case "fa-home":
+                    return "\xf015";
+                case "fa-hospital-o":
+                    return "\xf0f8";
+                case "fa-hotel":
+                    return "\xf236";
+                case "fa-hourglass":
+                    return "\xf254";
+                case "fa-hourglass-1":
+                    return "\xf251";
+                case "fa-hourglass-2":
+                    return "\xf252";
+                case "fa-hourglass-3":
+                    return "\xf253";
+                case "fa-hourglass-end":
+                    return "\xf253";
+                case "fa-hourglass-half":
+                    return "\xf252";
+                case "fa-hourglass-o":
+                    return "\xf250";
+                case "fa-hourglass-start":
+                    return "\xf251";
+                case "fa-houzz":
+                    return "\xf27c";
+                case "fa-html5":
+                    return "\xf13b";
+                case "fa-i-cursor":
+                    return "\xf246";
+                case "fa-id-badge":
+                    return "\xf2c1";
+                case "fa-id-card":
+                    return "\xf2c2";
+                case "fa-id-card-o":
+                    return "\xf2c3";
+                case "fa-ils":
+                    return "\xf20b";
+                case "fa-image":
+                    return "\xf03e";
+                case "fa-imdb":
+                    return "\xf2d8";
+                case "fa-inbox":
+                    return "\xf01c";
+                case "fa-indent":
+                    return "\xf03c";
+                case "fa-industry":
+                    return "\xf275";
+                case "fa-info":
+                    return "\xf129";
+                case "fa-info-circle":
+                    return "\xf05a";
+                case "fa-inr":
+                    return "\xf156";
+                case "fa-instagram":
+                    return "\xf16d";
+                case "fa-institution":
+                    return "\xf19c";
+                case "fa-internet-explorer":
+                    return "\xf26b";
+                case "fa-intersex":
+                    return "\xf224";
+                case "fa-ioxhost":
+                    return "\xf208";
+                case "fa-italic":
+                    return "\xf033";
+                case "fa-joomla":
+                    return "\xf1aa";
+                case "fa-jpy":
+                    return "\xf157";
+                case "fa-jsfiddle":
+                    return "\xf1cc";
+                case "fa-key":
+                    return "\xf084";
+                case "fa-keyboard-o":
+                    return "\xf11c";
+                case "fa-krw":
+                    return "\xf159";
+                case "fa-language":
+                    return "\xf1ab";
+                case "fa-laptop":
+                    return "\xf109";
+                case "fa-lastfm":
+                    return "\xf202";
+                case "fa-lastfm-square":
+                    return "\xf203";
+                case "fa-leaf":
+                    return "\xf06c";
+                case "fa-leanpub":
+                    return "\xf212";
+                case "fa-legal":
+                    return "\xf0e3";
+                case "fa-lemon-o":
+                    return "\xf094";
+                case "fa-level-down":
+                    return "\xf149";
+                case "fa-level-up":
+                    return "\xf148";
+                case "fa-life-bouy":
+                    return "\xf1cd";
+                case "fa-life-buoy":
+                    return "\xf1cd";
+                case "fa-life-ring":
+                    return "\xf1cd";
+                case "fa-life-saver":
+                    return "\xf1cd";
+                case "fa-lightbulb-o":
+                    return "\xf0eb";
+                case "fa-line-chart":
+                    return "\xf201";
+                case "fa-link":
+                    return "\xf0c1";
+                case "fa-linkedin":
+                    return "\xf0e1";
+                case "fa-linkedin-square":
+                    return "\xf08c";
+                case "fa-linode":
+                    return "\xf2b8";
+                case "fa-linux":
+                    return "\xf17c";
+                case "fa-list":
+                    return "\xf03a";
+                case "fa-list-alt":
+                    return "\xf022";
+                case "fa-list-ol":
+                    return "\xf0cb";
+                case "fa-list-ul":
+                    return "\xf0ca";
+                case "fa-location-arrow":
+                    return "\xf124";
+                case "fa-lock":
+                    return "\xf023";
+                case "fa-long-arrow-down":
+                    return "\xf175";
+                case "fa-long-arrow-left":
+                    return "\xf177";
+                case "fa-long-arrow-right":
+                    return "\xf178";
+                case "fa-long-arrow-up":
+                    return "\xf176";
+                case "fa-low-vision":
+                    return "\xf2a8";
+                case "fa-magic":
+                    return "\xf0d0";
+                case "fa-magnet":
+                    return "\xf076";
+                case "fa-mail-forward":
+                    return "\xf064";
+                case "fa-mail-reply":
+                    return "\xf112";
+                case "fa-mail-reply-all":
+                    return "\xf122";
+                case "fa-male":
+                    return "\xf183";
+                case "fa-map":
+                    return "\xf279";
+                case "fa-map-marker":
+                    return "\xf041";
+                case "fa-map-o":
+                    return "\xf278";
+                case "fa-map-pin":
+                    return "\xf276";
+                case "fa-map-signs":
+                    return "\xf277";
+                case "fa-mars":
+                    return "\xf222";
+                case "fa-mars-double":
+                    return "\xf227";
+                case "fa-mars-stroke":
+                    return "\xf229";
+                case "fa-mars-stroke-h":
+                    return "\xf22b";
+                case "fa-mars-stroke-v":
+                    return "\xf22a";
+                case "fa-maxcdn":
+                    return "\xf136";
+                case "fa-meanpath":
+                    return "\xf20c";
+                case "fa-medium":
+                    return "\xf23a";
+                case "fa-medkit":
+                    return "\xf0fa";
+                case "fa-meetup":
+                    return "\xf2e0";
+                case "fa-meh-o":
+                    return "\xf11a";
+                case "fa-mercury":
+                    return "\xf223";
+                case "fa-microchip":
+                    return "\xf2db";
+                case "fa-microphone":
+                    return "\xf130";
+                case "fa-microphone-slash":
+                    return "\xf131";
+                case "fa-minus":
+                    return "\xf068";
+                case "fa-minus-circle":
+                    return "\xf056";
+                case "fa-minus-square":
+                    return "\xf146";
+                case "fa-minus-square-o":
+                    return "\xf147";
+                case "fa-mixcloud":
+                    return "\xf289";
+                case "fa-mobile":
+                    return "\xf10b";
+                case "fa-mobile-phone":
+                    return "\xf10b";
+                case "fa-modx":
+                    return "\xf285";
+                case "fa-money":
+                    return "\xf0d6";
+                case "fa-moon-o":
+                    return "\xf186";
+                case "fa-mortar-board":
+                    return "\xf19d";
+                case "fa-motorcycle":
+                    return "\xf21c";
+                case "fa-mouse-pointer":
+                    return "\xf245";
+                case "fa-music":
+                    return "\xf001";
+                case "fa-navicon":
+                    return "\xf0c9";
+                case "fa-neuter":
+                    return "\xf22c";
+                case "fa-newspaper-o":
+                    return "\xf1ea";
+                case "fa-object-group":
+                    return "\xf247";
+                case "fa-object-ungroup":
+                    return "\xf248";
+                case "fa-odnoklassniki":
+                    return "\xf263";
+                case "fa-odnoklassniki-square":
+                    return "\xf264";
+                case "fa-opencart":
+                    return "\xf23d";
+                case "fa-openid":
+                    return "\xf19b";
+                case "fa-opera":
+                    return "\xf26a";
+                case "fa-optin-monster":
+                    return "\xf23c";
+                case "fa-outdent":
+                    return "\xf03b";
+                case "fa-pagelines":
+                    return "\xf18c";
+                case "fa-paint-brush":
+                    return "\xf1fc";
+                case "fa-paper-plane":
+                    return "\xf1d8";
+                case "fa-paper-plane-o":
+                    return "\xf1d9";
+                case "fa-paperclip":
+                    return "\xf0c6";
+                case "fa-paragraph":
+                    return "\xf1dd";
+                case "fa-paste":
+                    return "\xf0ea";
+                case "fa-pause":
+                    return "\xf04c";
+                case "fa-pause-circle":
+                    return "\xf28b";
+                case "fa-pause-circle-o":
+                    return "\xf28c";
+                case "fa-paw":
+                    return "\xf1b0";
+                case "fa-paypal":
+                    return "\xf1ed";
+                case "fa-pencil":
+                    return "\xf040";
+                case "fa-pencil-square":
+                    return "\xf14b";
+                case "fa-pencil-square-o":
+                    return "\xf044";
+                case "fa-percent":
+                    return "\xf295";
+                case "fa-phone":
+                    return "\xf095";
+                case "fa-phone-square":
+                    return "\xf098";
+                case "fa-photo":
+                    return "\xf03e";
+                case "fa-picture-o":
+                    return "\xf03e";
+                case "fa-pie-chart":
+                    return "\xf200";
+                case "fa-pied-piper":
+                    return "\xf2ae";
+                case "fa-pied-piper-alt":
+                    return "\xf1a8";
+                case "fa-pied-piper-pp":
+                    return "\xf1a7";
+                case "fa-pinterest":
+                    return "\xf0d2";
+                case "fa-pinterest-p":
+                    return "\xf231";
+                case "fa-pinterest-square":
+                    return "\xf0d3";
+                case "fa-plane":
+                    return "\xf072";
+                case "fa-play":
+                    return "\xf04b";
+                case "fa-play-circle":
+                    return "\xf144";
+                case "fa-play-circle-o":
+                    return "\xf01d";
+                case "fa-plug":
+                    return "\xf1e6";
+                case "fa-plus":
+                    return "\xf067";
+                case "fa-plus-circle":
+                    return "\xf055";
+                case "fa-plus-square":
+                    return "\xf0fe";
+                case "fa-plus-square-o":
+                    return "\xf196";
+                case "fa-podcast":
+                    return "\xf2ce";
+                case "fa-power-off":
+                    return "\xf011";
+                case "fa-print":
+                    return "\xf02f";
+                case "fa-product-hunt":
+                    return "\xf288";
+                case "fa-puzzle-piece":
+                    return "\xf12e";
+                case "fa-qq":
+                    return "\xf1d6";
+                case "fa-qrcode":
+                    return "\xf029";
+                case "fa-question":
+                    return "\xf128";
+                case "fa-question-circle":
+                    return "\xf059";
+                case "fa-question-circle-o":
+                    return "\xf29c";
+                case "fa-quora":
+                    return "\xf2c4";
+                case "fa-quote-left":
+                    return "\xf10d";
+                case "fa-quote-right":
+                    return "\xf10e";
+                case "fa-ra":
+                    return "\xf1d0";
+                case "fa-random":
+                    return "\xf074";
+                case "fa-ravelry":
+                    return "\xf2d9";
+                case "fa-rebel":
+                    return "\xf1d0";
+                case "fa-recycle":
+                    return "\xf1b8";
+                case "fa-reddit":
+                    return "\xf1a1";
+                case "fa-reddit-alien":
+                    return "\xf281";
+                case "fa-reddit-square":
+                    return "\xf1a2";
+                case "fa-refresh":
+                    return "\xf021";
+                case "fa-registered":
+                    return "\xf25d";
+                case "fa-remove":
+                    return "\xf00d";
+                case "fa-renren":
+                    return "\xf18b";
+                case "fa-reorder":
+                    return "\xf0c9";
+                case "fa-repeat":
+                    return "\xf01e";
+                case "fa-reply":
+                    return "\xf112";
+                case "fa-reply-all":
+                    return "\xf122";
+                case "fa-resistance":
+                    return "\xf1d0";
+                case "fa-retweet":
+                    return "\xf079";
+                case "fa-rmb":
+                    return "\xf157";
+                case "fa-road":
+                    return "\xf018";
+                case "fa-rocket":
+                    return "\xf135";
+                case "fa-rotate-left":
+                    return "\xf0e2";
+                case "fa-rotate-right":
+                    return "\xf01e";
+                case "fa-rouble":
+                    return "\xf158";
+                case "fa-rss":
+                    return "\xf09e";
+                case "fa-rss-square":
+                    return "\xf143";
+                case "fa-rub":
+                    return "\xf158";
+                case "fa-ruble":
+                    return "\xf158";
+                case "fa-rupee":
+                    return "\xf156";
+                case "fa-s15":
+                    return "\xf2cd";
+                case "fa-safari":
+                    return "\xf267";
+                case "fa-save":
+                    return "\xf0c7";
+                case "fa-scissors":
+                    return "\xf0c4";
+                case "fa-scribd":
+                    return "\xf28a";
+                case "fa-search":
+                    return "\xf002";
+                case "fa-search-minus":
+                    return "\xf010";
+                case "fa-search-plus":
+                    return "\xf00e";
+                case "fa-sellsy":
+                    return "\xf213";
+                case "fa-send":
+                    return "\xf1d8";
+                case "fa-send-o":
+                    return "\xf1d9";
+                case "fa-server":
+                    return "\xf233";
+                case "fa-share":
+                    return "\xf064";
+                case "fa-share-alt":
+                    return "\xf1e0";
+                case "fa-share-alt-square":
+                    return "\xf1e1";
+                case "fa-share-square":
+                    return "\xf14d";
+                case "fa-share-square-o":
+                    return "\xf045";
+                case "fa-shekel":
+                    return "\xf20b";
+                case "fa-sheqel":
+                    return "\xf20b";
+                case "fa-shield":
+                    return "\xf132";
+                case "fa-ship":
+                    return "\xf21a";
+                case "fa-shirtsinbulk":
+                    return "\xf214";
+                case "fa-shopping-bag":
+                    return "\xf290";
+                case "fa-shopping-basket":
+                    return "\xf291";
+                case "fa-shopping-cart":
+                    return "\xf07a";
+                case "fa-shower":
+                    return "\xf2cc";
+                case "fa-sign-in":
+                    return "\xf090";
+                case "fa-sign-language":
+                    return "\xf2a7";
+                case "fa-sign-out":
+                    return "\xf08b";
+                case "fa-signal":
+                    return "\xf012";
+                case "fa-signing":
+                    return "\xf2a7";
+                case "fa-simplybuilt":
+                    return "\xf215";
+                case "fa-sitemap":
+                    return "\xf0e8";
+                case "fa-skyatlas":
+                    return "\xf216";
+                case "fa-skype":
+                    return "\xf17e";
+                case "fa-slack":
+                    return "\xf198";
+                case "fa-sliders":
+                    return "\xf1de";
+                case "fa-slideshare":
+                    return "\xf1e7";
+                case "fa-smile-o":
+                    return "\xf118";
+                case "fa-snapchat":
+                    return "\xf2ab";
+                case "fa-snapchat-ghost":
+                    return "\xf2ac";
+                case "fa-snapchat-square":
+                    return "\xf2ad";
+                case "fa-snowflake-o":
+                    return "\xf2dc";
+                case "fa-soccer-ball-o":
+                    return "\xf1e3";
+                case "fa-sort":
+                    return "\xf0dc";
+                case "fa-sort-alpha-asc":
+                    return "\xf15d";
+                case "fa-sort-alpha-desc":
+                    return "\xf15e";
+                case "fa-sort-amount-asc":
+                    return "\xf160";
+                case "fa-sort-amount-desc":
+                    return "\xf161";
+                case "fa-sort-asc":
+                    return "\xf0de";
+                case "fa-sort-desc":
+                    return "\xf0dd";
+                case "fa-sort-down":
+                    return "\xf0dd";
+                case "fa-sort-numeric-asc":
+                    return "\xf162";
+                case "fa-sort-numeric-desc":
+                    return "\xf163";
+                case "fa-sort-up":
+                    return "\xf0de";
+                case "fa-soundcloud":
+                    return "\xf1be";
+                case "fa-space-shuttle":
+                    return "\xf197";
+                case "fa-spinner":
+                    return "\xf110";
+                case "fa-spoon":
+                    return "\xf1b1";
+                case "fa-spotify":
+                    return "\xf1bc";
+                case "fa-square":
+                    return "\xf0c8";
+                case "fa-square-o":
+                    return "\xf096";
+                case "fa-stack-exchange":
+                    return "\xf18d";
+                case "fa-stack-overflow":
+                    return "\xf16c";
+                case "fa-star":
+                    return "\xf005";
+                case "fa-star-half":
+                    return "\xf089";
+                case "fa-star-half-empty":
+                    return "\xf123";
+                case "fa-star-half-full":
+                    return "\xf123";
+                case "fa-star-half-o":
+                    return "\xf123";
+                case "fa-star-o":
+                    return "\xf006";
+                case "fa-steam":
+                    return "\xf1b6";
+                case "fa-steam-square":
+                    return "\xf1b7";
+                case "fa-step-backward":
+                    return "\xf048";
+                case "fa-step-forward":
+                    return "\xf051";
+                case "fa-stethoscope":
+                    return "\xf0f1";
+                case "fa-sticky-note":
+                    return "\xf249";
+                case "fa-sticky-note-o":
+                    return "\xf24a";
+                case "fa-stop":
+                    return "\xf04d";
+                case "fa-stop-circle":
+                    return "\xf28d";
+                case "fa-stop-circle-o":
+                    return "\xf28e";
+                case "fa-street-view":
+                    return "\xf21d";
+                case "fa-strikethrough":
+                    return "\xf0cc";
+                case "fa-stumbleupon":
+                    return "\xf1a4";
+                case "fa-stumbleupon-circle":
+                    return "\xf1a3";
+                case "fa-subscript":
+                    return "\xf12c";
+                case "fa-subway":
+                    return "\xf239";
+                case "fa-suitcase":
+                    return "\xf0f2";
+                case "fa-sun-o":
+                    return "\xf185";
+                case "fa-superpowers":
+                    return "\xf2dd";
+                case "fa-superscript":
+                    return "\xf12b";
+                case "fa-support":
+                    return "\xf1cd";
+                case "fa-table":
+                    return "\xf0ce";
+                case "fa-tablet":
+                    return "\xf10a";
+                case "fa-tachometer":
+                    return "\xf0e4";
+                case "fa-tag":
+                    return "\xf02b";
+                case "fa-tags":
+                    return "\xf02c";
+                case "fa-tasks":
+                    return "\xf0ae";
+                case "fa-taxi":
+                    return "\xf1ba";
+                case "fa-telegram":
+                    return "\xf2c6";
+                case "fa-television":
+                    return "\xf26c";
+                case "fa-tencent-weibo":
+                    return "\xf1d5";
+                case "fa-terminal":
+                    return "\xf120";
+                case "fa-text-height":
+                    return "\xf034";
+                case "fa-text-width":
+                    return "\xf035";
+                case "fa-th":
+                    return "\xf00a";
+                case "fa-th-large":
+                    return "\xf009";
+                case "fa-th-list":
+                    return "\xf00b";
+                case "fa-themeisle":
+                    return "\xf2b2";
+                case "fa-thermometer":
+                    return "\xf2c7";
+                case "fa-thermometer-0":
+                    return "\xf2cb";
+                case "fa-thermometer-1":
+                    return "\xf2ca";
+                case "fa-thermometer-2":
+                    return "\xf2c9";
+                case "fa-thermometer-3":
+                    return "\xf2c8";
+                case "fa-thermometer-4":
+                    return "\xf2c7";
+                case "fa-thermometer-empty":
+                    return "\xf2cb";
+                case "fa-thermometer-full":
+                    return "\xf2c7";
+                case "fa-thermometer-half":
+                    return "\xf2c9";
+                case "fa-thermometer-quarter":
+                    return "\xf2ca";
+                case "fa-thermometer-three-quarters":
+                    return "\xf2c8";
+                case "fa-thumb-tack":
+                    return "\xf08d";
+                case "fa-thumbs-down":
+                    return "\xf165";
+                case "fa-thumbs-o-down":
+                    return "\xf088";
+                case "fa-thumbs-o-up":
+                    return "\xf087";
+                case "fa-thumbs-up":
+                    return "\xf164";
+                case "fa-ticket":
+                    return "\xf145";
+                case "fa-times":
+                    return "\xf00d";
+                case "fa-times-circle":
+                    return "\xf057";
+                case "fa-times-circle-o":
+                    return "\xf05c";
+                case "fa-times-rectangle":
+                    return "\xf2d3";
+                case "fa-times-rectangle-o":
+                    return "\xf2d4";
+                case "fa-tint":
+                    return "\xf043";
+                case "fa-toggle-down":
+                    return "\xf150";
+                case "fa-toggle-left":
+                    return "\xf191";
+                case "fa-toggle-off":
+                    return "\xf204";
+                case "fa-toggle-on":
+                    return "\xf205";
+                case "fa-toggle-right":
+                    return "\xf152";
+                case "fa-toggle-up":
+                    return "\xf151";
+                case "fa-trademark":
+                    return "\xf25c";
+                case "fa-train":
+                    return "\xf238";
+                case "fa-transgender":
+                    return "\xf224";
+                case "fa-transgender-alt":
+                    return "\xf225";
+                case "fa-trash":
+                    return "\xf1f8";
+                case "fa-trash-o":
+                    return "\xf014";
+                case "fa-tree":
+                    return "\xf1bb";
+                case "fa-trello":
+                    return "\xf181";
+                case "fa-tripadvisor":
+                    return "\xf262";
+                case "fa-trophy":
+                    return "\xf091";
+                case "fa-truck":
+                    return "\xf0d1";
+                case "fa-try":
+                    return "\xf195";
+                case "fa-tty":
+                    return "\xf1e4";
+                case "fa-tumblr":
+                    return "\xf173";
+                case "fa-tumblr-square":
+                    return "\xf174";
+                case "fa-turkish-lira":
+                    return "\xf195";
+                case "fa-tv":
+                    return "\xf26c";
+                case "fa-twitch":
+                    return "\xf1e8";
+                case "fa-twitter":
+                    return "\xf099";
+                case "fa-twitter-square":
+                    return "\xf081";
+                case "fa-umbrella":
+                    return "\xf0e9";
+                case "fa-underline":
+                    return "\xf0cd";
+                case "fa-undo":
+                    return "\xf0e2";
+                case "fa-universal-access":
+                    return "\xf29a";
+                case "fa-university":
+                    return "\xf19c";
+                case "fa-unlink":
+                    return "\xf127";
+                case "fa-unlock":
+                    return "\xf09c";
+                case "fa-unlock-alt":
+                    return "\xf13e";
+                case "fa-unsorted":
+                    return "\xf0dc";
+                case "fa-upload":
+                    return "\xf093";
+                case "fa-usb":
+                    return "\xf287";
+                case "fa-usd":
+                    return "\xf155";
+                case "fa-user":
+                    return "\xf007";
+                case "fa-user-circle":
+                    return "\xf2bd";
+                case "fa-user-circle-o":
+                    return "\xf2be";
+                case "fa-user-md":
+                    return "\xf0f0";
+                case "fa-user-o":
+                    return "\xf2c0";
+                case "fa-user-plus":
+                    return "\xf234";
+                case "fa-user-secret":
+                    return "\xf21b";
+                case "fa-user-times":
+                    return "\xf235";
+                case "-users":
+                    return "\xf0c0";
+                case "-vcard":
+                    return "\xf2bb";
+                case "fa-vcard-o":
+                    return "\xf2bc";
+                case "fa-venus":
+                    return "\xf221";
+                case "fa-venus-double":
+                    return "\xf226";
+                case "fa-venus-mars":
+                    return "\xf228";
+                case "fa-viacoin":
+                    return "\xf237";
+                case "fa-viadeo":
+                    return "\xf2a9";
+                case "fa-viadeo-square":
+                    return "\xf2aa";
+                case "fa-video-camera":
+                    return "\xf03d";
+                case "fa-vimeo":
+                    return "\xf27d";
+                case "fa-vimeo-square":
+                    return "\xf194";
+                case "fa-vine":
+                    return "\xf1ca";
+                case "fa-vk":
+                    return "\xf189";
+                case "fa-volume-control-phone":
+                    return "\xf2a0";
+                case "fa-volume-down":
+                    return "\xf027";
+                case "fa-volume-off":
+                    return "\xf026";
+                case "fa-volume-up":
+                    return "\xf028";
+                case "fa-warning":
+                    return "\xf071";
+                case "fa-wechat":
+                    return "\xf1d7";
+                case "fa-weibo":
+                    return "\xf18a";
+                case "fa-weixin":
+                    return "\xf1d7";
+                case "fa-whatsapp":
+                    return "\xf232";
+                case "fa-wheelchair":
+                    return "\xf193";
+                case "fa-wheelchair-alt":
+                    return "\xf29b";
+                case "fa-wifi":
+                    return "\xf1eb";
+                case "fa-wikipedia-w":
+                    return "\xf266";
+                case "fa-window-close":
+                    return "\xf2d3";
+                case "fa-window-close-o":
+                    return "\xf2d4";
+                case "fa-window-maximize":
+                    return "\xf2d0";
+                case "fa-window-minimize":
+                    return "\xf2d1";
+                case "fa-window-restore":
+                    return "\xf2d2";
+                case "fa-windows":
+                    return "\xf17a";
+                case "fa-won":
+                    return "\xf159";
+                case "fa-wordpress":
+                    return "\xf19a";
+                case "fa-wpbeginner":
+                    return "\xf297";
+                case "fa-wpexplorer":
+                    return "\xf2de";
+                case "fa-wpforms":
+                    return "\xf298";
+                case "fa-wrench":
+                    return "\xf0ad";
+                case "fa-xing":
+                    return "\xf168";
+                case "fa-xing-square":
+                    return "\xf169";
+                case "fa-y-combinator":
+                    return "\xf23b";
+                case "fa-y-combinator-square":
+                    return "\xf1d4";
+                case "fa-yahoo":
+                    return "\xf19e";
+                case "fa-yc":
+                    return "\xf23b";
+                case "fa-yc-square":
+                    return "\xf1d4";
+                case "fa-yelp":
+                    return "\xf1e9";
+                case "fa-yen":
+                    return "\xf157";
+                case "fa-yoast":
+                    return "\xf2b1";
+                case "fa-youtube":
+                    return "\xf167";
+                case "fa-youtube-play":
+                    return "\xf16a";
+                case "fa-youtube-square":
+                    return "\xf166";
+                default:
+                    return "";
+            }
+        }
+    }
+}
