+++ NetShare.sln
@@ -1,13 +1,15 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.32929.386
+# Visual Studio Version 17
+VisualStudioVersion = 17.8.34408.163
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetShare", "NetShare\NetShare.csproj", "{881B8D8B-5D73-4EBE-878A-5043F817C880}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infrastracture", "Infrastracture\Infrastracture.csproj", "{56DB0554-FF1D-450B-90DC-75B74721113C}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "Domain\Domain.csproj", "{8F4CCD0B-F5A4-400C-AAFA-6E976EEC13DD}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FontAwesome", "FontAwesome\FontAwesome.csproj", "{83B4FE88-1621-4583-B320-05BE8F38C7C4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +29,10 @@ Global
 		{8F4CCD0B-F5A4-400C-AAFA-6E976EEC13DD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8F4CCD0B-F5A4-400C-AAFA-6E976EEC13DD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8F4CCD0B-F5A4-400C-AAFA-6E976EEC13DD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{83B4FE88-1621-4583-B320-05BE8F38C7C4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{83B4FE88-1621-4583-B320-05BE8F38C7C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{83B4FE88-1621-4583-B320-05BE8F38C7C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{83B4FE88-1621-4583-B320-05BE8F38C7C4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
